+++ src/en/twebor/mobmasks/maskeffects/MaskChicken.java
@@ -1,0 +1,7 @@
+package en.twebor.mobmasks.maskeffects;
+
+public class MaskChicken{
+
+    public MaskChicken() {
+    }
+}
