--- src/en/twebor/mobmasks/masklogic/MaskChicken.java
@@ -1,9 +1,0 @@
-package en.twebor.mobmasks.masklogic;
-
-import org.bukkit.plugin.java.JavaPlugin;
-
-public class MaskChicken{
-
-    public MaskChicken() {
-    }
-}
